requirements.txt
asgiref==3.8.1
build==1.2.1
click==8.1.7
Django==5.0.4
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.15.1
packaging==24.0
pip-tools==7.4.1
pyproject_hooks==1.0.0
setuptools==69.2.0
sqlparse==0.4.4
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.20240311
typing_extensions==4.11.0
wheel==0.43.0
